+++ lib/views/home/sub_views/adaptive_view/adaptive_list/transactions/list_view_transaction_splits.dart
@@ -60,6 +60,7 @@ class _ListViewTransactionSplitsState extends State<ListViewTransactionSplits> {
             list: rows,
             selectedItemIds: ValueNotifier<List<int>>([]),
             onSelectionChanged: (int _) {},
+            scrollController: ScrollController(),
           ),
         ),
       ],