+++ lib/views/home/sub_views/view_investments/view_investments_details_panels.dart
@@ -79,6 +79,7 @@ extension ViewInvestmentsDetailsPanels on ViewInvestmentsState {
     final SelectionController selectionController = Get.put(SelectionController());
     return ListViewTransactions(
       key: Key(instance.uniqueId.toString()),
+      listController: Get.find<ListControllerInfoPanel>(),
       columnsToInclude: <Field>[
         Transaction.fields.getFieldByName(columnIdDate),
         Transaction.fields.getFieldByName(columnIdAccount),